# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HackForge Hackathon Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    put:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectTitle:
          type: string
        description:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, projectTitle, createdAt]
    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
        projectTitle:
          type: string
        description:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
      required: [name, projectTitle]
...